component created for Drag and drop functionality - setState passed through props
